Change the snapshot directory format, removing `__Snapshots__`
Instead of adding a directory called `__Snapshots__`, as is the default
in the SnapshotTesting library, this change just appends a `.snapshots`
to the name of the current file, and uses that as the directory for
the snapshots.

This makes it more readable when adding the snapshot directories to
Xcode as (unlinked) groups.